Fix the workbench test by disabling various settings that should be disabled for the test pod.
The previous method of calling set and passing . separated names for subkeys does not work.
This method creates the proper objects for all disabled objects and combines them into the
overrideDictionary .Values to ensure everything is overridden properly. mergeOverwrite is used
to ensure that we forcefully disable these settings as they should NEVER be set within a test pod.