Add support for generating a ServiceMonitor CRD for RSPM.
This is required in clusters that use the Prometheus Operator instead of
a regular Prometheus server. It's also a very common feature of public
Helm charts for applications with Prometheus metrics.

Note that there are many other fields in this CRD [0] that I have not
yet exposed via values.yml. In researching comparable Helm charts, I
found that these fields were very inconsistently exposed, so I decided
to start with the smallest useful subset instead.

[0]: https://github.com/prometheus-operator/prometheus-operator/blob/main/Documentation/api.md#endpoint

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>