rspm: Use a ClusterIP service by default instead of a NodePort.
This is a breaking change.

ClusterIP is the Kubernetes default, exposing the service in-cluster
only. This is a secure, well-understood default.

For external users, there are a few options:

* Use an Ingress. We support this already and it works with ClusterIP.

* Use a LoadBalancer. This can be expensive so it makes a poor default.

* Use a NodePort and some kind of external load balancing solution.

It seems like the last route (which is the current default) is much more
uncommon than the first two, so I'd advocate for this change.

In addition:

* NodePorts are a very limited cluster resource, so we should avoid
  consuming them if we can avoid it.

* Defaulting to NodePort is a poor security choice, especially if we
  have users that don't understand the consequences.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>