Remove the requirement for privileged RSPM containers by default.
Instead, set a more granular security context: enable only the necessary
capabilities (which includes SYS_ADMIN at this time) and disable
AppArmor via annotation.

Note that the default security context still assumes that RSPM is
running as root. When that is the case, RSPM runs an R version check on
startup that cannot be disabled even if `Git.AllowUnsandboxedGitBuilds =
true`, hence the need to include these settings unconditionally.

When not running as root, no additional capabilities are required and
seccomp and AppArmor must be disabled only to run sandboxed Git builds,
which the user can opt out of.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>