Add a Helm hook for migrating RSPM's shared storage permissions.
This should automatically fix the root vs. non-root file permissions
issue. Unfortunately, there is no way to know whether the upgrade we are
about to perform necessitates this migration, so the Job always runs
regardless of whether it is needed.

A new 'enableMigrations' setting allows this hook to be disabled; in the
future when we think users will no longer need it this can become false
by default.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>