remove job-json-overrides from templates when useTemplates is set
Before, when useTemplates was set, we still inject `job-json-overrides` into the launcher.kubernetes.profiles.conf output and set up the requisite mounts. launcher.kubernetes.profiles.conf is always necessary, but job-json-overrides conflicts with useTemplates and causes launcher to warn (or perhaps someday error). As a result, we now pivot properly and pull out the job-json-overrides injections, config, and mounts.