ensure that job-json-overrides is off when templating is on
This is complicated because the old mechanism for `job-json-overrides` has still not been removed, so we technically have three options available. The default / recommendation for the chart is still the "newer" `job-json-overrides`. We have updated the "older" `jobJsonOverridesFiles / profiles` to be deprecated. When profiles are defined in `config.server`, we ignore the "new" profiles setup. However, we now warn in `NOTES.txt`. If `useTemplates` is true, then we cut out the "default" overrides but continue using "profiles" as expected (using `config.profiles` by default and a fall-back to "as-provided" `config.server`). We really need to remove some of these layers to make the configuration generation clearer and easier to understand :scream: