add an adhoc implementation of command for the workbench job template
This is important for a support article about
customizing the entrypoint script for Workbench
sessions. Ultimately, we probably need to expose
this sort of behavior as a product feature, similar to
Connect's "supervisor script." In the meantime,
this should hopefully suffice as a workaround for
what was previously done with `job-json-overrides`.