Allow Launcher to set K8s API URL and Token from pod
This change is required to support the rotation of the service account
token that is the default behavior in newer versions of Kubernetes.

The `prestart.bash` script for Connect was exporting
- `KUBERNETES_AUTH_TOKEN`
- `KUBERNETES_API_URL`

These were used by Connect to set entries in the configuration file.

When the token that was mounted in the container at startup would
expire, all successive requests would fail with a `401` because it was
using the auth token that was explicitly written to
`launcher.kubernetes.conf`.